Creating a "Chatbot" to interact with users in Chinese.
If users enter "能告诉我天津有什么电影吗"（Can you tell me what movies are playing in Tianjin?）, the Chatbot will find the citynames and reply with a list of the recent release movies.

Based on the list the Chatbot gave, users will ask whether the movie is worth wathing, the Chatbot will give the users the rating, the introduction and the original website url of the movie, if they are interested to click in for more information .

And the Chatbot can also be trained with your own chat-history. (just put yml files in the Corpus repository.)